% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_dat.R
\docType{data}
\name{area_dat}
\alias{area_dat}
\title{Area level dataset}
\format{
A tibble with X rows and Y variables:
\describe{
\item{varname}{short desc}
}
}
\source{
\url{https://github.com/USDAForestService/FIESTA}
}
\usage{
area_dat
}
\description{
Area level dataset
}
\keyword{datasets}
