% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_unit_level.R
\name{basic_unit_level}
\alias{basic_unit_level}
\title{The basic unit-level small area estimation model.}
\usage{
basic_unit_level(
  mode = "regression",
  small_areas = NULL,
  explanatory_means = NULL,
  small_area_size = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible value is "regression".}

\item{areas}{The small areas.}
}
\description{
The basic unit-level small area estimation model.
}
