% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hba.R
\name{hba}
\alias{hba}
\title{The area-level hierarchical Bayesian model}
\usage{
hba(data, formula, variance)
}
\arguments{
\item{data}{the data}

\item{formula}{formula for fixed effects}

\item{variance}{within-area variance}
}
\description{
The area-level hierarchical Bayesian model
}
