% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbsae_area.R
\name{hbsae_area}
\alias{hbsae_area}
\title{The area-level hierarchical Bayesian model}
\usage{
hbsae_area(data, formula, small_area, pop_data, post_strat_data)
}
\arguments{
\item{data}{the data}

\item{formula}{formula for fixed effects}

\item{small_area}{character vector of length one. random effects term.}

\item{pop_data}{population data for the explanatory variable.}

\item{post_strat_data}{y variable}
}
\description{
The area-level hierarchical Bayesian model
}
