% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbsae_cov.R
\name{hbsae_cov}
\alias{hbsae_cov}
\title{Coefficient of Variation for hbsae}
\usage{
hbsae_cov(data)
}
\arguments{
\item{data}{an object created from hbsae_*}
}
\description{
Coefficient of Variation for hbsae
}
