% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbsae_unit.R
\name{hbsae_unit}
\alias{hbsae_unit}
\title{The unit-level hierarchical Bayesian model}
\usage{
hbsae_unit(data, formula, small_area, pop_data)
}
\arguments{
\item{data}{the data}

\item{formula}{formula for fixed effects}

\item{small_area}{character vector of length one. random effects term.}

\item{pop_data}{population data for the explanatory variable.}
}
\description{
The unit-level hierarchical Bayesian model
}
