% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_area.R
\name{sae_area}
\alias{sae_area}
\title{The area-level frequentist (EBLUP) model}
\usage{
sae_area(data, formula, small_area, pop_data, post_strat_data)
}
\arguments{
\item{data}{the data}

\item{formula}{formula for fixed effects}

\item{small_area}{character vector of length one. random effects term.}

\item{pop_data}{population data for the explanatory variable.}

\item{post_strat_data}{y variable}
}
\description{
The area-level frequentist (EBLUP) model
}
