% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_unit_cov.R
\name{sae_unit_cov}
\alias{sae_unit_cov}
\title{Bootstrapped coefficient of variation for the unit-level EBLUP}
\usage{
sae_unit_cov(data, formula, small_area, pop_data, B = 100)
}
\arguments{
\item{data}{the data}

\item{formula}{formula for fixed effects}

\item{small_area}{character vector of length one. random effects term.}

\item{pop_data}{population data for the explanatory variable.}

\item{B}{number of bootstrap samples}
}
\description{
Bootstrapped coefficient of variation for the unit-level EBLUP
}
