% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_area_level.R
\name{basic_area_level}
\alias{basic_area_level}
\title{The basic area-level small area estimation model.}
\usage{
basic_area_level(mode = "regression", within_area_variance = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible value is "regression".}

\item{within_area_variance}{The within-area variance for the small area model.}
}
\description{
The basic area-level small area estimation model.
}
